Add files via upload
For my capstone project in data analytics, I wanted to look at the impact of climate change on the housing market. I started with flood insurance claims in Florida, a state known for being hit by severe weather more as time goes on. This Tableau packaged workbook is a little funkier than my others because I used a lot of floating elements in my story to get everything "just so" when I screen-shared my presentation to my class. But I the data and dashboards should still be good.